fix: update g-svg to solve some todo tests.
