+++ stories/svg/svg.stories.tsx
@@ -3,6 +3,7 @@ import { storiesOf } from '@storybook/react';
 import React from 'react';
 import Grid from './component/grid';
 import ContextMenu from './component/contextMenu';
+import Tooltip from './component/tooltip';
 
 export default { title: 'SVG' };
 
@@ -12,4 +13,7 @@ storiesOf('SVG', module)
 ))
 .add('context menu', () => (
   <ContextMenu />
+))
+.add('tooltip', () => (
+  <Tooltip />
 ))
